I walked into a town a little worse for wear.
The path has not been kind to me over the past few days.
I'd worn the soles to my shoes, and taken a tumble on my way into town.

Determined to make the best of my situation I joined some of the
locals for a meal at their tavern. The owner, a cobler, helped
"worn shoes".sub(/^\w+/,'new')

I'll be leaving tomorrow in lighter spirits, and on comforted feet.