The fog lifts. I've found truth.
Not so much universal truth and enlightenment, but the truth of my
situation. Where I had dispared in the past I should rejoice!

Each hardship has lead me through some meaningful task.
The most menial of tasks has lead me to understand something more
of the world in which I reside.

I count myself lucky. That is the truth.